Postman tests

+++ src/main/java/nl/projectautoplanner/projectautoplannerwebapi/Services/GebruikerService.java
@@ -2,9 +2,8 @@ package nl.projectautoplanner.projectautoplannerwebapi.Services;
 
 import nl.projectautoplanner.projectautoplannerwebapi.DomainModels.Gebruiker;
 import nl.projectautoplanner.projectautoplannerwebapi.Repositories.GebruikerRepository;
-import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Service;
-import java.util.List;
 
 @Service
 public class GebruikerService {
@@ -15,21 +14,20 @@ public class GebruikerService {
         this.gebruikerRepository = gebruikerRepository;
     }
 
-    public List<Gebruiker> getAllGebruikers() {
-        return gebruikerRepository.findAll();
-    }
-
-    public Gebruiker getGebruikerById(Long id) {
-        return gebruikerRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Gebruiker niet gevonden"));
-    }
-
     public Gebruiker saveGebruiker(Gebruiker gebruiker) {
         return gebruikerRepository.save(gebruiker);
     }
 
-    public void deleteGebruiker(Long id) {
-        gebruikerRepository.deleteById(id);
+    public void deleteGebruiker(Long id, Jwt jwt) {
+        String ingelogdeNaam = jwt.getClaimAsString("preferred_username");
+        Gebruiker teVerwijderen = gebruikerRepository.findById(id).orElseThrow();
+        boolean isAdmin = jwt.getClaimAsMap("resource_access")
+                .get("ProjectautoPlanner").toString().contains("ADMIN");
+        if (isAdmin || teVerwijderen.getGebruikersnaam().equalsIgnoreCase(ingelogdeNaam)) {
+            gebruikerRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Je mag alleen je eigen account verwijderen!");
+        }
     }
 
     public Gebruiker updateGebruiker(Long id, Gebruiker nieuweData) {
@@ -41,5 +39,16 @@ public class GebruikerService {
             bestaandeGebruiker.setAdres(nieuweData.getAdres());
             return gebruikerRepository.save(bestaandeGebruiker);
         }).orElseThrow(() -> new RuntimeException("Gebruiker niet gevonden"));
+    }
+    public Gebruiker getGebruikerByGebruikersnaam(String gebruikersnaam) {
+        return gebruikerRepository.findByGebruikersnaamIgnoreCase(gebruikersnaam)
+                .orElseThrow(() -> new RuntimeException("Gebruiker niet gevonden."));
+    }
+
+    public Gebruiker updateRolByGebruikersnaam(String gebruikersnaam, Gebruiker.GebruikerRol rolEnum) {
+        Gebruiker gebruiker = gebruikerRepository.findByGebruikersnaamIgnoreCase(gebruikersnaam)
+                .orElseThrow(() -> new RuntimeException("Gebruiker met naam " + gebruikersnaam + " niet gevonden"));
+        gebruiker.setRol(rolEnum);
+        return gebruikerRepository.save(gebruiker);
     }
 }
